requirements.txt
flask==3.0.0
ultralytics==8.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.0
Pillow==10.0.1
werkzeug==3.0.1
importlib-metadata; python_version < "3.8"